support daemon/query

+++ aports/local/timeskew/timeskew.sh
@@ -1,27 +1,37 @@
 #!/usr/bin/env bash
 PIDFILE="/run/timeskew.pid"
 
-_daemon() {
-  local pid cur lt rt delta
-  pid="$$"
-  echo "$pid" > $PIDFILE
+_run() {
+  local pid cur lt rt delta is_daemon
+  is_daemon=0
+  if [ "$1" == "--daemon" ]; then
+    is_daemon=1
+  fi
+  if [ "$is_daemon" -eq 1 ]; then
+    pid="$$"
+    echo "$pid" > $PIDFILE
+  fi
   while : ; do
-    if [ ! -e "$PIDFILE" ]; then
-      return
+    if [ "$is_daemon" -eq 1 ]; then
+      if [ ! -e "$PIDFILE" ]; then
+        return
+      fi
+      cur=$(cat "$PIDFILE")
+      if [ "$cur" != "$pid" ]; then
+        return
+      fi
+      sleep 10
     fi
-    cur=$(cat "$PIDFILE")
-    if [ "$cur" != "$pid" ]; then
-      return
-    fi
-    sleep 10
     rt=$(curl --silent "http://router.voidedtech.com/time")
     if [ -n "$rt" ] && [ "$rt" -eq "$rt" ] 2>/dev/null; then
       lt=$(date +%s)
       delta=0
     else
       echo "failed to get remote time" | logger -t timeskew
-      sleep 5
-      continue
+      if [ "$is_daemon" ]; then
+        sleep 5
+        continue
+      fi
     fi
     if [ "$lt" -gt "$rt" ]; then
       delta=$((lt-rt))
@@ -30,14 +40,27 @@ _daemon() {
         delta=$((rt-lt))
       fi
     fi
-    if [ "$delta" -gt 10 ]; then
-      {
-        echo "delta detected, exceeds threshold: $delta (local: $lt, remote: $rt)"
-        pkill chronyd
-        rc-service chronyd restart
-      } 2>&1 | logger -t timeskew
+    if [ "$is_daemon" -eq 1 ]; then
+      if [ "$delta" -gt 10 ]; then
+        {
+          echo "delta detected, exceeds threshold: $delta (local: $lt, remote: $rt)"
+          pkill chronyd
+          rc-service chronyd restart
+        } 2>&1 | logger -t timeskew
+      fi
+    else
+      echo "delta: $delta, remote: $rt, local: $lt"
+      return
     fi
   done
 }
 
-_daemon &
+if [ -n "$1" ]; then
+  if [ "$1" == "--daemon" ]; then
+    _run --daemon &
+    exit 0
+  fi
+  echo "unknown argument: $1"
+  exit 1
+fi
+_run
